Add bench and some cleaning
