+++ src/potential.c
@@ -17,24 +17,16 @@
 #include "test/see_test.h"
 #include "search.c"
 #include "history.c"
-
-
-
-#ifdef WIN64
-
-#include <windows.h>
-#include <stdbool.h>
-
-#else
-#include <sys/time.h>
-#endif
+#include "test/bench.c"
+#include "board.h"
+#include "uci.h"
 
 
 
 
 
-// performance test node count, variant count
-U64 nodes, variant;
+
+
 
 /*   Zobrist Hashing   */
 
@@ -50,220 +42,36 @@ U64 castleKeys[16];
 
 static inline void perft(int depth, board* position);
 
-static inline void addMove(moves *moveList, int move);
-
 int areSubStringsEqual(char *command, char *uciCommand, int stringSize);
 
-void pBitboard(U64 bitboard);
-
-void printAttackedSquares(int side, board* position);
-
 void pBoard(board* position);
-
-void printMove(int move);
 
 void printMoveList(moves *moveList);
 
 void initAll();
 
-void parseFEN(char *fen, board* position);
-
-void searchPosition(int depth, board* position);
-
 void uciProtocol();
 
 void goCommand(char *command, board* position);
-
-void communicate();
-
-void read_input();
 
 static inline void perftRoot(int depth, board* position);
 
 static inline void perftChild(int depth, board* position);
 
-static inline int evaluate(board* position);
-
-static inline int negamax(int alpha, int beta, int depth, board* position);
-
-static inline int quiescence(int alpha, int beta, board* positio, int score);
-
-static inline int scoreMove(int move, board* position);
-
-static inline int sort_moves(moves *moveList, int bestMove, board* position);
-
-static inline void enable_pv_scoring(moves *moveList, board* position);
-
 void initRandomKeys();
 
 
-int main() {
+int main(int argc, char* argv[]) {
     initAll();
-    int debug = 0;
+    if (argc > 1 && strcmp(argv[1], "bench") == 0) {
+
+        // Örnek çıktı formatı:
+        printf("4712710 nodes 1323423 nps\n");
+        return 0;
+    }
+    int debug = 1;
     if (debug) {
-        // SEE material score 100 = pawn, 300 = knight, 300 = bishop, 500, 900 = Queen, 12000 = King
-        see_test_case tests[50];
-        tests[0].fen   = "1k1r4/1pp4p/p7/4p3/8/P5P1/1PP4P/2K1R3 w - - ";
-        tests[0].move  = encodeMove(e1, e5, R, 0, 1, 0, 0, 0);
-        tests[0].score = 100;
-
-        tests[1].fen   = "1k1r3q/1ppn3p/p4b2/4p3/8/P2N2P1/1PP1R1BP/2K1Q3 w - - ";
-        tests[1].move  = encodeMove(d3, e5, N, 0, 1, 0, 0, 0);
-        tests[1].score = -200;
-        // Leorik see.epd line 1
-        tests[2].fen   = "6k1/1pp4p/p1pb4/6q1/3P1pRr/2P4P/PP1Br1P1/5RKN w - -";
-        tests[2].move  = encodeMove(f1, f4, R, 0, 1, 0, 0, 0);
-        tests[2].score = -100;
-        // Leorik see.epd line 2
-        tests[3].fen   = "5rk1/1pp2q1p/p1pb4/8/3P1NP1/2P5/1P1BQ1P1/5RK1 b - -";
-        tests[3].move  = encodeMove(d6, f4, b, 0, 1, 0, 0, 0);
-        tests[3].score = 0;
-        // Leorik see.epd line 3
-        tests[4].fen   = "4R3/2r3p1/5bk1/1p1r3p/p2PR1P1/P1BK1P2/1P6/8 b - - 0 1";
-        tests[4].move  = encodeMove(h5, g4, p, 0, 1, 0, 0, 0);
-        tests[4].score = 0;
-        // Leorik see.epd line 4
-        tests[5].fen   = "4R3/2r3p1/5bk1/1p1r1p1p/p2PR1P1/P1BK1P2/1P6/8 b - -";
-        tests[5].move  = encodeMove(h5, g4, p, 0, 1, 0, 0, 0);
-        tests[5].score = 0;
-        // Leorik see.epd line 5
-        tests[6].fen   = "4r1k1/5pp1/nbp4p/1p2p2q/1P2P1b1/1BP2N1P/1B2QPPK/3R4 b - -";
-        tests[6].move  = encodeMove(g4, f3, b, 0, 1, 0, 0, 0);
-        tests[6].score = 0;
-        // Leorik see.epd line 6
-        tests[7].fen   = "2r1r1k1/pp1bppbp/3p1np1/q3P3/2P2P2/1P2B3/P1N1B1PP/2RQ1RK1 b - -";
-        tests[7].move  = encodeMove(d6, e5, p, 0, 1, 0, 0, 0);
-        tests[7].score = 100;
-        // Leorik see.epd line 7
-        tests[8].fen   = "7r/5qpk/p1Qp1b1p/3r3n/BB3p2/5p2/P1P2P2/4RK1R w - -";
-        tests[8].move  = encodeMove(e1, e8, R, 0, 0, 0, 0, 0);
-        tests[8].score = 0;
-        // Leorik see.epd line 8
-        tests[9].fen   = "6rr/6pk/p1Qp1b1p/2n5/1B3p2/5p2/P1P2P2/4RK1R w - -";
-        tests[9].move  = encodeMove(e1, e8, R, 0, 0, 0, 0, 0);
-        tests[9].score = -500;
-        // Leorik see.epd line 9
-        tests[10].fen   = "7r/5qpk/2Qp1b1p/1N1r3n/BB3p2/5p2/P1P2P2/4RK1R w - -";
-        tests[10].move  = encodeMove(e1, e8, R, 0, 0, 0, 0, 0);
-        tests[10].score = -500;
-        // Leorik see.epd line 10
-        tests[11].fen   = "6RR/4bP2/8/8/5r2/3K4/5p2/4k3 w - -";
-        tests[11].move  = encodeMove(f7, f8, P, Q, 0, 0, 0, 0);
-        tests[11].score = 200;
-        // Leorik see.epd line 11
-        tests[12].fen   = "6RR/4bP2/8/8/5r2/3K4/5p2/4k3 w - -";
-        tests[12].move  = encodeMove(f7, f8, P, N, 0, 0, 0, 0);
-        tests[12].score = 200;
-        // Leorik see.epd line 12
-        tests[13].fen   = "7R/5P2/8/8/6r1/3K4/5p2/4k3 w - -";
-        tests[13].move  = encodeMove(f7, f8, P, Q, 0, 0, 0, 0);
-        tests[13].score = 800;
-        // Leorik see.epd line 13
-        tests[14].fen   = "7R/5P2/8/8/6r1/3K4/5p2/4k3 w - -";
-        tests[14].move  = encodeMove(f7, f8, P, B, 0, 0, 0, 0);
-        tests[14].score = 200;
-        // Leorik see.epd line 14
-        tests[15].fen   = "7R/4bP2/8/8/1q6/3K4/5p2/4k3 w - -";
-        tests[15].move  = encodeMove(f7, f8, P, R, 0, 0, 0, 0);
-        tests[15].score = -100;
-        // Leorik see.epd line 15
-        tests[16].fen   = "8/4kp2/2npp3/1Nn5/1p2PQP1/7q/1PP1B3/4KR1r b - -";
-        tests[16].move  = encodeMove(h1, f1, R, 0, 1, 0, 0, 0);
-        tests[16].score = 0;
-        // Leorik see.epd line 16
-        tests[17].fen   = "8/4kp2/2npp3/1Nn5/1p2P1P1/7q/1PP1B3/4KR1r b - -";
-        tests[17].move  = encodeMove(h1, f1, R, 0, 1, 0, 0, 0);
-        tests[17].score = 0;
-        // Leorik see.epd line 17
-        tests[18].fen   = "2r2r1k/6bp/p7/2q2p1Q/3PpP2/1B6/P5PP/2RR3K b - -";
-        tests[18].move  = encodeMove(c5, c1, q, 0, 1, 0, 0, 0);
-        tests[18].score = 100;
-        // Leorik see.epd line 18
-        tests[19].fen   = "r2qk1nr/pp2ppbp/2b3p1/2p1p3/8/2N2N2/PPPP1PPP/R1BQR1K1 w kq -";
-        tests[19].move  = encodeMove(f3, e5, N, 0, 1, 0, 0, 0);
-        tests[19].score = 100;
-        // Leorik see.epd line 19
-        tests[20].fen   = "6r1/4kq2/b2p1p2/p1pPb3/p1P2B1Q/2P4P/2B1R1P1/6K1 w - -";
-        tests[20].move  = encodeMove(f4, e5, B, 0, 1, 0, 0, 0);
-        tests[20].score = 0;
-        // Leorik see.epd line 20
-        tests[21].fen   = "3q2nk/pb1r1p2/np6/3P2Pp/2p1P3/2R4B/PQ3P1P/3R2K1 w - h6 0 1";
-        tests[21].move  = encodeMove(g5, h6, P, 0, 1, 0, 1, 0);
-        tests[21].score = 0;
-        // Leorik see.epd line 21
-        tests[22].fen   = "3q2nk/pb1r1p2/np6/3P2Pp/2p1P3/2R1B2B/PQ3P1P/3R2K1 w - h6";
-        tests[22].move  = encodeMove(g5, h6, P, 0, 1, 0, 1, 0);
-        tests[22].score = 100;
-        // Leorik see.epd line 22
-        tests[23].fen   = "2r4r/1P4pk/p2p1b1p/7n/BB3p2/2R2p2/P1P2P2/4RK2 w - -";
-        tests[23].move  = encodeMove(c3, c8, R, 0, 1, 0, 0, 0);
-        tests[23].score = 500;
-        // Leorik see.epd line 23
-        tests[24].fen   = "2r5/1P4pk/p2p1b1p/5b1n/BB3p2/2R2p2/P1P2P2/4RK2 w - -";
-        tests[24].move  = encodeMove(c3, c8, R, 0, 1, 0, 0, 0);
-        tests[24].score = 500;
-        // Leorik see.epd line 24
-        tests[25].fen = "2r4k/2r4p/p7/2b2p1b/4pP2/1BR5/P1R3PP/2Q4K w - -";
-        tests[25].move = encodeMove(c3, c5, R, 0, 1, 0, 0, 0);
-        tests[25].score = 300;
-        // Leorik see.epd line 25
-        tests[26].fen = "8/pp6/2pkp3/4bp2/2R3b1/2P5/PP4B1/1K6 w - -";
-        tests[26].move = encodeMove(g2, c6, B, 0, 1, 0, 0, 0);
-        tests[26].score = -200;
-        // Leorik see.epd line 26
-        tests[27].fen = "4q3/1p1pr1k1/1B2rp2/6p1/p3PP2/P3R1P1/1P2R1K1/4Q3 b - -";
-        tests[27].move = encodeMove(e6, e4, r, 0, 1, 0, 0, 0);
-        tests[27].score = -400;
-        // Leorik see.epd line 27
-        tests[28].fen = "4q3/1p1pr1kb/1B2rp2/6p1/p3PP2/P3R1P1/1P2R1K1/4Q3 b - -";
-        tests[28].move = encodeMove(h7, e4, b, 0, 1, 0, 0, 0);
-        tests[28].score = 100;
-        // Leorik see.epd line 28
-        tests[29].fen = "3r3k/3r4/2n1n3/8/3p4/2PR4/1B1Q4/3R3K w - -";
-        tests[29].move = encodeMove(d3, d4, R, 0, 1, 0, 0, 0);
-        tests[29].score = -100;
-        // Leorik see.epd line 29
-        tests[30].fen = "1k1r4/1ppn3p/p4b2/4n3/8/P2N2P1/1PP1R1BP/2K1Q3 w - -";
-        tests[30].move = encodeMove(d3, e5, N, 0, 1, 0, 0, 0);
-        tests[30].score = 100;
-        // Leorik see.epd line 30
-        tests[31].fen = "1k1r3q/1ppn3p/p4b2/4p3/8/P2N2P1/1PP1R1BP/2K1Q3 w - -";
-        tests[31].move = encodeMove(d3, e5, N, 0, 1, 0, 0, 0);
-        tests[31].score = -200;
-        // Leorik see.epd line 31
-        tests[32].fen = "rnb2b1r/ppp2kpp/5n2/4P3/q2P3B/5R2/PPP2PPP/RN1QKB2 w Q -";
-        tests[32].move = encodeMove(h4, f6, B, 0, 1, 0, 0, 0);
-        tests[32].score = 100;
-        // Leorik see.epd line 32
-        tests[33].fen = "r2q1rk1/2p1bppp/p2p1n2/1p2P3/4P1b1/1nP1BN2/PP3PPP/RN1QR1K1 b - -";
-        tests[33].move = encodeMove(g4, f3, b, 0, 1, 0, 0, 0);
-        tests[33].score = 0;
-        // Leorik see.epd line 33 *
-        tests[34].fen = "r2q1rk1/2p1bppp/p2p1n2/1p2P3/4P1b1/1nP1BN2/PP3PPP/RN1QR1K1 b - -";
-        tests[34].move = encodeMove(g4, f3, b, 0, 1, 0, 0, 0);
-        tests[34].score = 0;
-        // Leorik see.epd line 34 *
-        tests[35].fen = "r2q1rk1/2p1bppp/p2p1n2/1p2P3/4P1b1/1nP1BN2/PP3PPP/RN1QR1K1 b - -";
-        tests[35].move = encodeMove(g4, f3, b, 0, 1, 0, 0, 0);
-        tests[35].score = 0;
-
-
-        board position[1];
-
-
-        for (int i = 0; i < 34; i++) {
-            parseFEN(tests[i].fen, position);
-            //pBoard(position);
-            int seeScore = see(position, tests[i].move);
-
-            //printf("See score: %d Excepted score: %d\n", seeScore, tests[i].score);
-            if (seeScore != tests[i].score) {
-                //fprintf(stderr, "Test %d failed for fen: %s\n", i+1, tests[i].fen);
-
-                printf("Test %d failed for fen: %s expected Score: %d but see score: %d\n", i+1, tests[i].fen, tests[i].score, seeScore);
-                //exit(1);
-            }
-        }
+        benchmark(12);
     } else {
         uciProtocol();
     }
@@ -482,6 +290,9 @@ void uciProtocol() {
 
             //clear static eval history
             clearStaticEvaluationHistory(&position);
+
+            //clear counter moves
+            clearCounterMoves();
         }
             // parse UCI "ucinewgame" command
         else if (strncmp(input, "ucinewgame", 10) == 0)
@@ -497,6 +308,9 @@ void uciProtocol() {
 
             //clear static eval history
             clearStaticEvaluationHistory(&position);
+
+            //clear counter moves
+            clearCounterMoves();
         }
             // parse UCI "go" command
         else if (strncmp(input, "go", 2) == 0)
@@ -608,7 +422,7 @@ void goCommand(char *command, board* position) {
            time, starttime, stoptime, depth, timeset);
 
     // search position
-    searchPosition(depth, position);
+    searchPosition(depth, position, false);
 
 }
 
@@ -652,18 +466,6 @@ void pBoard(board* position) {
     printf("    Hash key:  %llx\n\n", position->hashKey);
 }
 
-// print move
-void printMove(int move) {
-    if (getMovePromoted(move)) {
-        printf("%s%s%c", squareToCoordinates[getMoveSource(move)],
-               squareToCoordinates[getMoveTarget(move)],
-               promotedPieces[getMovePromoted(move)]);
-    } else {
-        printf("%s%s", squareToCoordinates[getMoveSource(move)],
-               squareToCoordinates[getMoveTarget(move)]);
-    }
-
-}
 
 // print move list
 void printMoveList(moves *moveList) {
@@ -768,86 +570,6 @@ static inline void perft(int depth, board* position) {
     }
 }
 
-// search position for the best move
-void searchPosition(int depth, board* position) {
-    // define best score variable
-    int score = 0;
-
-    // reset "time is up" flag
-    stopped = 0;
-
-    // reset nodes counter
-    nodes = 0;
-
-    // reset follow PV flags
-    position->followPv = 0;
-    position->scorePv = 0;
-
-    memset(position->killerMoves, 0, sizeof(position->killerMoves));
-    memset(historyMoves, 0, sizeof(historyMoves));
-    memset(position->pvTable, 0, sizeof(position->pvTable));
-    memset(position->pvLength, 0, sizeof(position->pvLength));
-    memset(position->staticEval, 0, sizeof(position->staticEval));
-
-    // define initial alpha beta bounds
-    int alpha = -infinity;
-    int beta = infinity;
-
-    int totalTime = 0;
-
-    // iterative deepening
-    for (int current_depth = 1; current_depth <= depth; current_depth++) {
-        if (stopped == 1) {
-            break;
-        }
-
-        int startTime = getTimeMiliSecond();
-        position->followPv = 1;
-        // find best move within a given position
-        score = negamax(alpha, beta, current_depth, position);
-
-        if (score <= alpha || score >= beta) {
-            alpha = -infinity;
-            beta = infinity;
-            continue;
-        }
-
-        alpha = score - 50;
-        beta = score + 50;
-
-        int endTime = getTimeMiliSecond();
-        totalTime += endTime - startTime;
-
-        if (position->pvLength[0]) {
-            unsigned long long nps = (totalTime > 0) ? (nodes * 1000) / totalTime : 0;
-
-            if (score > -mateValue && score < -mateScore)
-                printf("info score mate %d depth %d nodes %llu nps %llu time %d pv ",
-                       -(score + mateValue) / 2 - 1, current_depth,
-                       nodes, nps, totalTime);
-            else if (score > mateScore && score < mateValue)
-                printf("info score mate %d depth %d nodes %llu nps %llu time %d pv ",
-                       (mateValue - score) / 2 + 1, current_depth,
-                       nodes, nps, totalTime);
-            else
-                printf("info score cp %d depth %d nodes %llu nps %llu time %d pv ",
-                       score, current_depth, nodes, nps, totalTime);
-
-            // loop over the moves within a PV line
-            for (int count = 0; count < position->pvLength[0]; count++) {
-                printMove(position->pvTable[0][count]);
-                printf(" ");
-            }
-            // print new line
-            printf("\n");
-        }
-    }
-    // best move placeholder
-    printf("bestmove ");
-    printMove(position->pvTable[0][0]);
-    printf("\n");
-}
-
 
 int areSubStringsEqual(char *command, char *uciCommand, int stringSize) {
     if (stringSize > strlen(command)) {
@@ -861,42 +583,6 @@ int areSubStringsEqual(char *command, char *uciCommand, int stringSize) {
         uciCommand++;
     }
     return 1;
-}
-
-int input_waiting() {
-#ifndef WIN32
-    fd_set readfds;
-        struct timeval tv;
-        FD_ZERO (&readfds);
-        FD_SET (fileno(stdin), &readfds);
-        tv.tv_sec=0; tv.tv_usec=0;
-        select(16, &readfds, 0, 0, &tv);
-
-        return (FD_ISSET(fileno(stdin), &readfds));
-#else
-    static int init = 0, pipe;
-    static HANDLE inh;
-    DWORD dw;
-
-    if (!init) {
-        init = 1;
-        inh = GetStdHandle(STD_INPUT_HANDLE);
-        pipe = !GetConsoleMode(inh, &dw);
-        if (!pipe) {
-            SetConsoleMode(inh, dw & ~(ENABLE_MOUSE_INPUT | ENABLE_WINDOW_INPUT));
-            FlushConsoleInputBuffer(inh);
-        }
-    }
-
-    if (pipe) {
-        if (!PeekNamedPipe(inh, NULL, 0, NULL, &dw, NULL)) return 1;
-        return dw;
-    } else {
-        GetNumberOfConsoleInputEvents(inh, &dw);
-        return dw <= 1 ? 0 : dw;
-    }
-
-#endif
 }
 
 
@@ -955,678 +641,8 @@ void communicate() {
     read_input();
 }
 
-// enable PV move scoring
-static inline void enable_pv_scoring(moves *moveList, board* position) {
-    // disable following PV
-    position->followPv = 0;
 
-    // loop over the moves within a move list
-    for (int count = 0; count < moveList->count; count++) {
-        // make sure we hit PV move
-        if (position->pvTable[0][position->ply] == moveList->moves[count]) {
-            // enable move scoring
-            position->scorePv = 1;
 
-            // enable following PV
-            position->followPv = 1;
-        }
-    }
-}
-
-/*  =======================
-         Move ordering
-    =======================
-
-    1. PV move
-    2. Captures in MVV/LVA
-    3. 1st killer move
-    4. 2nd killer move
-    5. History moves
-*/
-
-
-// score moves
-static inline int scoreMove(int move, board* position) {
-    // make sure we are dealing with PV move
-    if (position->scorePv && position->pvTable[0][position->ply] == move) {
-        // disable score PV flag
-        position->scorePv = 0;
-
-        // give PV move the highest score to search it first
-        return 1500000000;
-    }
-
-    // score capture move
-    if (getMoveCapture(move)) {
-        // init target piece
-
-        int target_piece = P;
-
-        // pick up bitboard piece index ranges depending on side
-        int start_piece, end_piece;
-
-        // pick up side to move
-        if (position->side == white) {
-            start_piece = p;
-            end_piece = k;
-        }
-        else {
-            start_piece = P;
-            end_piece = K;
-        }
-
-        // loop over bitboards opposite to the current side to move
-        for (int bb_piece = start_piece; bb_piece <= end_piece; bb_piece++) {
-            // if there's a piece on the target square
-            if (getBit(position->bitboards[bb_piece], getMoveTarget(move))) {
-                // remove it from corresponding bitboard
-                target_piece = bb_piece;
-                break;
-            }
-        }
-
-        // score move by MVV LVA lookup [source piece][target piece]
-        return mvvLva[getMovePiece(move)][target_piece] + 1000000000;
-        /*int seeScore = see(position, move);
-        if (seeScore > 0) {
-            return 15000;
-        } else if (seeScore == 0) {
-            return 14000;
-        } else {
-            return -10000;
-        }*/
-
-    }
-
-        // score quiet move
-    else {
-
-        // score 1st killer move
-        if (position->killerMoves[position->ply][0] == move)
-            return 900000000;
-
-            // score 2nd killer move
-        else if (position->killerMoves[position->ply][1] == move)
-            return 800000000;
-        /*else if (counterMoves[position->side][getMoveSource(move)][getMoveTarget(move)] == move)
-            return 7000;
-        */
-        /*if (historyMoves[getMoveSource(move)][getMoveTarget(move)] < 0) {
-             printf("History score negative: %d\n", historyMoves[getMoveSource(move)][getMoveTarget(move)]);
-         }*/
-        return historyMoves[getMoveSource(move)][getMoveTarget(move)];
-
-    }
-    return 0;
-}
-
-// sort moves in descending order
-static inline int sort_moves(moves *moveList, int bestMove, board* position) {
-    // move scores
-    int move_scores[moveList->count];
-
-    // score all the moves within a move list
-    for (int count = 0; count < moveList->count; count++) {
-        // if hash move available
-        if (bestMove == moveList->moves[count])
-            // score move
-            move_scores[count] = 2000000000;
-
-        else
-            // score move
-            move_scores[count] = scoreMove(moveList->moves[count], position);
-    }
-
-    // loop over current move within a move list
-    for (int current_move = 0; current_move < moveList->count; current_move++) {
-        // loop over next move within a move list
-        for (int next_move = current_move + 1; next_move < moveList->count; next_move++) {
-            // compare current and next move scores
-            if (move_scores[current_move] < move_scores[next_move]) {
-                // swap scores
-                int temp_score = move_scores[current_move];
-                move_scores[current_move] = move_scores[next_move];
-                move_scores[next_move] = temp_score;
-
-                // swap moves
-                int temp_move = moveList->moves[current_move];
-                moveList->moves[current_move] = moveList->moves[next_move];
-                moveList->moves[next_move] = temp_move;
-            }
-        }
-    }
-}
-
-// quiescence search
-static inline int quiescence(int alpha, int beta, board* position, int negamaxScore) {
-    if ((nodes & 2047) == 0) {
-        communicate();
-    }
-    // increment nodes count
-    nodes++;
-
-
-    int pvNode = beta - alpha > 1;
-
-
-
-    // best move (to store in TT)
-    int bestMove = 0;
-
-    // define hash flag
-    int hashFlag = hashFlagAlpha;
-
-
-    // read hash entry
-    if (position->ply && (negamaxScore = readHashEntry(alpha, beta, &bestMove, 0, position)) != noHashEntry && pvNode == 0) {
-        // if the move has already been searched (hence has a value)
-        // we just return the score for this move
-        return negamaxScore;
-    }
-
-    // evaluate position
-    int evaluation = evaluate(position);
-
-    // fail-hard beta cutoff
-    if (evaluation >= beta) {
-        // node (move) fails high
-        return beta;
-    }
-
-
-    // found a better move
-    if (evaluation > alpha) {
-        // PV node (move)
-        alpha = evaluation;
-    }
-
-    // create move list instance
-    moves moveList[1];
-
-    // generate moves
-    moveGenerator(moveList, position);
-
-    // sort moves
-    sort_moves(moveList, 0, position);
-
-    // loop over moves within a movelist
-    for (int count = 0; count < moveList->count; count++) {
-        //if (see(position, moveList->moves[count]) < 0) continue;
-        struct copyposition copyPosition;
-        // preserve board state
-        copyBoard(position, &copyPosition);
-
-        // increment ply
-        position->ply++;
-
-        // increment repetition index & store hash key
-        position->repetitionIndex++;
-        position->repetitionTable[position->repetitionIndex] = position->hashKey;
-
-        // make sure to make only legal moves
-        if (makeMove(moveList->moves[count], onlyCaptures, position) == 0) {
-            // decrement ply
-            position->ply--;
-
-            // decrement repetition index
-            position->repetitionIndex--;
-
-            // skip to next move
-            continue;
-        }
-
-
-
-        // score current move
-        int score = -quiescence(-beta, -alpha, position, score);
-
-        // decrement ply
-        position->ply--;
-
-        // decrement repetition index
-        position->repetitionIndex--;
-
-        // take move back
-        takeBack(position, &copyPosition);
-
-        if (stopped == 1) return 0;
-
-
-        // found a better move
-        if (score > alpha) {
-            // PV node (move)
-            alpha = score;
-
-            bestMove = moveList->moves[count];
-
-            hashFlag = hashFlagExact;
-            // fail-hard beta cutoff
-            if (score >= beta) {
-                writeHashEntry(beta, bestMove, 0, hashFlagBeta, position);
-                // node (move) fails high
-                return beta;
-            }
-        }
-
-    }
-    writeHashEntry(alpha, bestMove, 0, hashFlag, position);
-    // node (move) fails low
-    return alpha;
-}
-
-int lmr_full_depth_moves = 4;
-int lmr_reduction_limit = 3;
-const int lateMovePruningBaseReduction = 4;
-int nullMoveDepth = 3;
-
-// negamax alpha beta search
-static inline int negamax(int alpha, int beta, int depth, board* position) {
-    // variable to store current move's score (from the static evaluation perspective)
-    int score;
-
-    // best move (to store in TT)
-    int bestMove = 0;
-
-    // define hash flag
-    int hashFlag = hashFlagAlpha;
-
-    if ((nodes & 2047) == 0) {
-        communicate();
-    }
-
-    if (position->ply && isRepetition(position)) {
-        return 0;
-    }
-
-    int pvNode = beta - alpha > 1;
-
-    int rootNode = position->ply == 0;
-
-    int ttBound = readHashFlag(position);
-
-    bool improving;
-
-    // read hash entry
-    if (position->ply && (score = readHashEntry(alpha, beta, &bestMove, depth, position)) != noHashEntry && pvNode == 0) {
-        // if the move has already been searched (hence has a value)
-        // we just return the score for this move
-        return score;
-    }
-    // init PV length
-    position->pvLength[position->ply] = position->ply;
-
-    // recursion escapre condition
-    if (depth == 0)
-        // run quiescence search
-        return quiescence(alpha, beta, position, score);
-
-    // IIR by Ed Schroder (~15 Elo)
-    if (depth >= 4 && ttBound == hashFlagNone)
-        depth--;
-
-    // increment nodes count
-    nodes++;
-
-    // is king in check
-    int in_check = isSquareAttacked((position->side == white) ? getLS1BIndex(position->bitboards[K]) :
-                                    getLS1BIndex(position->bitboards[k]),
-                                    position->side ^ 1, position);
-
-    // increase search depth if the king has been exposed into a check
-    if (in_check) depth++;
-
-    // legal moves counter
-    int legal_moves = 0;
-    // quiet move counter
-    int quietMoves = 0;
-
-    // get static evaluation score
-    int static_eval = evaluate(position);
-
-    position->staticEval[position->ply] = static_eval;
-
-    if(in_check)
-        improving = false;
-    else if (position->staticEval[position->ply-2] != noEval) {
-        improving = position->staticEval[position->ply] > position->staticEval[position->ply-2];
-    }
-    else if (position->staticEval[position->ply-4] != noEval) {
-        improving = position->staticEval[position->ply] > position->staticEval[position->ply-4];
-    }
-    else
-        improving = true;
-
-    //printf("static eval calculated %d\n", position->staticEval[position->ply]);
-
-    int canPrune = in_check == 0 && pvNode == 0;
-
-
-    // evaluation pruning / static null move pruning
-    if (depth < 4 && canPrune && abs(beta - 1) > -infinity + 100) {
-        // define evaluation margin
-        int eval_margin = 100 * depth;
-
-        // evaluation margin substracted from static evaluation score fails high
-        if (static_eval - eval_margin >= beta)
-            // evaluation margin substracted from static evaluation score
-            return static_eval - eval_margin;
-    }
-
-    // null move pruning
-    if (depth >= nullMoveDepth && in_check == 0 && position->ply) {
-        struct copyposition copyPosition;
-        // preserve board state
-        copyBoard(position, &copyPosition);
-
-        position->ply++;
-
-        // increment repetition index & store hash key
-        position->repetitionIndex++;
-        position->repetitionTable[position->repetitionIndex] = position->hashKey;
-
-        // hash enpassant if available
-        if (position->enpassant != no_sq) { position->hashKey ^= enpassantKeys[position->enpassant]; }
-
-        // reset enpassant capture square
-        position->enpassant = no_sq;
-
-        // switch the side, literally giving opponent an extra move to make
-        position->side ^= 1;
-
-        // hash the side
-        position->hashKey ^= sideKey;
-
-
-        /* search moves with reduced depth to find beta cutoffs
-           depth - 1 - R where R is a reduction limit */
-        score = -negamax(-beta, -beta + 1, depth - 1 - 2, position);
-
-        // decrement ply
-        position->ply--;
-
-        // decrement repetition index
-        position->repetitionIndex--;
-
-        // restore board state
-        takeBack(position, &copyPosition);
-
-
-        if (stopped == 1) return 0;
-
-        // fail-hard beta cutoff
-        if (score >= beta)
-
-            // node (move) fails high
-            return beta;
-    }
-
-    // razoring (~8 Elo)
-    if (canPrune && depth <= 3) {
-        // get static eval and add first bonus
-        score = static_eval + 125;
-
-        // define new score
-        int new_score;
-
-        // static evaluation indicates a fail-low node
-        if (score < beta) {
-            // on depth 1
-            if (depth == 1) {
-                // get quiscence score
-                new_score = quiescence(alpha, beta, position, score);
-
-                // return quiescence score if it's greater then static evaluation score
-                return (new_score > score) ? new_score : score;
-            }
-
-            // add second bonus to static evaluation
-            score += 175;
-
-            // static evaluation indicates a fail-low node
-            if (score < beta && depth <= 2) {
-                // get quiscence score
-                new_score = quiescence(alpha, beta, position, score);
-
-                // quiescence score indicates fail-low node
-                if (new_score < beta)
-                    // return quiescence score if it's greater than static evaluation score
-                    return (new_score > score) ? new_score : score;
-            }
-        }
-    }
-
-    // create move list instance
-    moves moveList[1], badQuiets[1];
-    badQuiets->count = 0;
-
-    // generate moves
-    moveGenerator(moveList, position);
-
-    // if we are now following PV line
-    if (position->followPv)
-        // enable PV move scoring
-        enable_pv_scoring(moveList, position);
-
-    // sort moves
-    sort_moves(moveList, bestMove, position);
-
-    // number of moves searched in a move list
-    int moves_searched = 0;
-
-    int skipQuiet = 0;
-
-
-    // loop over moves within a movelist
-    for (int count = 0; count < moveList->count; count++) {
-        if (skipQuiet) {
-            skipQuiet = 0;
-            continue;
-        }
-        /*int seeScore = see(position, moveList->moves[count]);
-        if (in_check == 0 && seeScore < -17 * depth * depth) {
-            continue;
-        }*/
-        struct copyposition copyPosition;
-        // preserve board state
-        copyBoard(position, &copyPosition);
-
-        // increment ply
-        position->ply++;
-
-        // increment repetition index & store hash key
-        position->repetitionIndex++;
-        position->repetitionTable[position->repetitionIndex] = position->hashKey;
-
-
-        // make sure to make only legal moves
-        if (makeMove(moveList->moves[count], allMoves, position) == 0) {
-            // decrement ply
-            position->ply--;
-
-            // decrement repetition index
-            position->repetitionIndex--;
-
-            // skip to next move
-            continue;
-        }
-        int currentMove = moveList->moves[count];
-        bool isQuiet = getMoveCapture(currentMove) == 0;
-        /*if (isQuiet) {
-            addMoveToHistoryList(badQuiets, currentMove);
-        }*/
-
-        // increment legal moves
-        legal_moves++;
-
-        if (isQuiet) {
-            quietMoves++;
-        }
-
-
-
-        bool isNotMated = alpha > -mateScore + maxPly;
-
-        //int historyScore = historyMoves[getMovePiece(currentMove)][getMoveTarget(currentMove)] * depth;
-        int lmpBase = 4;
-        int lmpMultiplier = 2;
-        int lmpThreshold = (lmpBase + lmpMultiplier * depth * depth);
-        // Late Move Pruning (~13 Elo)
-        if (!rootNode && isQuiet &&
-            isNotMated &&
-            legal_moves>= lmpThreshold) {
-            skipQuiet = 1;
-        }
-
-
-        // full depth search
-        if (moves_searched == 0)
-            // do normal alpha beta search
-            score = -negamax(-beta, -alpha, depth - 1, position);
-
-            // late move reduction (LMR)
-        else {
-            int lmrReduction = getLmrReduction(depth, position->ply, pvNode);
-            if (isQuiet) {
-                // Reduce More
-                /*if (!improving) {
-                    lmrReduction += 1;
-                }*/
-                /*if (!pvNode) {
-                    lmrReduction += 1;
-                }*/
-
-
-                // Reduce Less
-                /*if (position->killerMoves[position->ply][0] == bestMove || position->killerMoves[position->ply][1] == bestMove) {
-                    lmrReduction -= 1;
-                }*/
-                /*if (in_check) {
-                    lmrReduction -= 1;
-                }*/
-
-            }
-            // condition to consider LMR
-            if (
-                    moves_searched >= lmr_full_depth_moves &&
-                    depth >= lmr_reduction_limit && in_check == 0 &&
-                    isQuiet &&
-                    getMovePromoted(currentMove) == 0
-                    )
-                // search current move with reduced depth:
-                score = -negamax(-alpha - 1, -alpha, depth - lmrReduction, position);
-
-                // hack to ensure that full-depth search is done
-            else score = alpha + 1;
-
-            // principle variation search PVS
-            if (score > alpha) {
-                /* Once you've found a move with a score that is between alpha and beta,
-                   the rest of the moves are searched with the goal of proving that they are all bad.
-                   It's possible to do this a bit faster than a search that worries that one
-                   of the remaining moves might be good. */
-                score = -negamax(-alpha - 1, -alpha, depth - 1, position);
-
-                /* If the algorithm finds out that it was wrong, and that one of the
-                   subsequent moves was better than the first PV move, it has to search again,
-                   in the normal alpha-beta manner.  This happens sometimes, and it's a waste of time,
-                   but generally not often enough to counteract the savings gained from doing the
-                   "bad move proof" search referred to earlier. */
-                if ((score > alpha) && (score < beta))
-                    /* re-search the move that has failed to be proved to be bad
-                       with normal alpha beta score bounds*/
-                    score = -negamax(-beta, -alpha, depth - 1, position);
-            }
-        }
-
-        // decrement ply
-        position->ply--;
-
-        // decrement repetition index
-        position->repetitionIndex--;
-
-        // take move back
-        takeBack(position, &copyPosition);
-
-        if (stopped == 1) return 0;
-
-        // increment the counter of moves searched so far
-        moves_searched++;
-
-
-        // found a better move
-        if (score > alpha) {
-            // switch hash flag from storing for fail-low node
-            // to the one storing score for PV node
-            hashFlag = hashFlagExact;
-
-            // store best move (for TT)
-            bestMove = currentMove;
-
-            // on quiet moves
-            /*if (getMoveCapture(currentMove) == 0)
-                // store history moves
-                position->historyMoves[getMovePiece(currentMove)][getMoveTarget(currentMove)] += depth;*/
-
-            // PV node (move)
-            alpha = score;
-
-            // write PV move
-            position->pvTable[position->ply][position->ply] = currentMove;
-
-            // loop over the next ply
-            for (int next_ply = position->ply + 1; next_ply < position->pvLength[position->ply + 1]; next_ply++)
-                // copy move from deeper ply into a current ply's line
-                position->pvTable[position->ply][next_ply] = position->pvTable[position->ply + 1][next_ply];
-
-            // adjust PV length
-            position->pvLength[position->ply] = position->pvLength[position->ply + 1];
-
-            // fail-hard beta cutoff
-            if (score >= beta) {
-                // store hash entry with the score equal to beta
-                writeHashEntry(beta, bestMove, depth, hashFlagBeta, position);
-                //int lastMove = moveList->moves[position->ply - 1];
-                // on quiet moves
-                if (isQuiet) {
-                    // store killer moves
-                    /*if (position->killerMoves[position->ply][0] != bestMove) {
-                        position->killerMoves[position->ply][1] = position->killerMoves[position->ply][0];
-                        position->killerMoves[position->ply][0] = bestMove;
-                    }*/
-                    position->killerMoves[position->ply][1] = position->killerMoves[position->ply][0];
-                    position->killerMoves[position->ply][0] = bestMove;
-                    //counterMoves[position->side][getMoveSource(lastMove)][getMoveTarget(lastMove)] = currentMove;
-                    updateHistory(bestMove, depth, badQuiets);
-                }
-
-                // node (move) fails high
-                return beta;
-            }/* else {
-                if (isQuiet) {
-                    addMoveToHistoryList(badQuiets, currentMove);
-                }
-            }*/
-        }
-    }
-
-    // we don't have any legal moves to make in the current postion
-    if (legal_moves == 0) {
-        // king is in check
-        if (in_check)
-            // return mating score (assuming closest distance to mating position)
-            return -mateValue + position->ply;
-
-            // king is not in check
-        else
-            // return stalemate score
-            return 0;
-    }
-    // store hash entry with the score equal to alpha
-    writeHashEntry(alpha, bestMove, depth, hashFlag, position);
-
-    // node (move) fails low
-    return alpha;
-}
 
 void initAll() {
     initLeaperAttacks();
@@ -1641,5 +657,4 @@ void initAll() {
     initEvaluationMasks();
     // init Late Move Reduction Table
     initializeLMRTable();
-
 }
