+++ src/test/bench.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "bit_manipulation.h"
+#include <stdio.h>
+#include "board.h"
+#include "search.h"
+#include "time.h"
+#include "evaluation.h"
+
+void benchmark();
