+++ src/time.h
@@ -2,10 +2,6 @@
 // Created by erena on 29.05.2024.
 //
 #pragma once
-#ifndef CHESSENGINE_TIME_H
-#define CHESSENGINE_TIME_H
-
-#endif //CHESSENGINE_TIME_H
 
 #ifdef WIN64
 
@@ -15,6 +11,33 @@
 #else
 #include <sys/time.h>
 #endif
+
+// UCI "movestogo" command moves counter
+extern int movestogo;
+
+// UCI "movetime" command time counter
+extern int movetime;
+
+// UCI "time" command holder (ms)
+extern int time;
+
+// UCI "inc" command's time increment holder
+extern int inc;
+
+// UCI "starttime" command time holder
+extern int starttime;
+
+// UCI "stoptime" command time holder
+extern int stoptime;
+
+// variable to flag time control availability
+extern int timeset;
+
+// variable to flag when the time is up
+extern int stopped;
+
+// exit from engine flag
+extern int quit;
 
 void resetTimeControl();
 int getTimeMiliSecond();
